backend/requirements.txt
asgiref==3.7.2
boto3==1.34.0
botocore==1.34.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.8
django-admin-datta==1.0.10
django-cors-headers==4.2.0
django-environ==0.11.2
django-extra-fields==3.0.2
django-filter==23.3
django-rest-swagger==2.2.0
django-s3-storage==0.14.0
django-webpack-loader==2.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.26.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.9.0
setuptools==69.0.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
